Refactor search feature
The search feature required loading all accounts into memory before searching.
This patch allows accounts to be processed one-by-one without loading them all
into memory. This allows more accounts to be processed.

Additionally, a mythril.cfg file is created in the ~/.mythril directory. This
file stores the path to the leveldb database. This file is automatically created
when the the ~/.mythril directory doesn't exist.

Mythril now checks to see if the leveldb database exists before searching. If it
does not exist, the user will receive an error.

Finally, the old persistence layer, ZODB requirements, and the --init-db
feature have been removed.